requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements-in.txt
#
appdirs==1.4.3
chromedriver-installer==0.0.6
click==6.7
packaging==16.8
py==1.4.33
pyparsing==2.2.0
pytest==3.0.7
python-dotenv==0.6.4
selenium==3.3.3
six==1.10.0

# The following packages are considered to be unsafe in a requirements file:
# setuptools                # via pytest
